🐛 Fix type cast error in payment JOIN query
- linked_student_id and student_id are TEXT in payments table
- s.id is BIGINT in students table
- Added ::bigint cast to fix type mismatch in JOIN
- p.linked_student_id::bigint = s.id
- p.student_id::bigint = s.id

+++ check-beatrisa-payment.sql
@@ -75,8 +75,8 @@ SELECT
   s.name as actual_student_name
 FROM payments p
 LEFT JOIN students s ON (
-  p.linked_student_id = s.id 
-  OR p.student_id = s.id
+  p.linked_student_id::bigint = s.id 
+  OR p.student_id::bigint = s.id
 )
 WHERE 
   s.name ILIKE '%beatrisa%arushanyan%'
